Implemented Mixer Controller to control levels, organized audio in unity
Added MixerController script
Added Audio objects in main menu and level 1 that have the MixerComponent script as well as most of the audio inside of it as child objects.
The important exceptions to this is when a sound is tied to a specific object that is placed elsewhere (like the fountain sound) in which case it is a script on the fountain object. Or the footstepController which must be on the player object that has the anim handler